Changed the default location for labscript files to C:\user_scripts\labscriptlib
Changed the default location for where to save experiment h5 files.

If you are on windows, and the becnas share is mounted, it will now detect the mount point (Z:\, etc), and build up a path for the h5 file.

This will be Z:\Experiments\<lab>\<labscript>\<year>-<month>\<day>
<Lab> is determined from the control server specified in labscript. If it is localhost, it looks up the PC name and chooses either "spinor_lab" or "dual_species_lab\<krb or narb>\"
<labscript> is the python file name without the .py extension

The default location for opening and creating global h5 files is then set to Z:\Experiments\<lab>\<labscript>

If you manually set the output directory after this, it will not revert when you select a new labscript file. To make it automatically adjust again, first manually set the dir to be the correct one following Z:\Experiments\<lab>\<labscript>\<year>-<month>\<day>

It will automatically advance the <day> when it rolls past midnight.

If you set the clock back in time, it will not update the folder as it rolls onto the next day, until it reaches the point in time when you set the clock back (this is because gobject timeouts don't update relative times, and so Run Manager has queued up a gobject timeout for 1s in the future, however in changing the clock back, you have moved that point further into the future). Daylight savings adjustments do not have any affect on this, so we are fine there!