Fixed bug where diffing globals files didn't work.
This was because the runmanager API had an line in it that was an attempt
at forward compatibility for Python 3, aliasing 'str' to the name 'unicode'.

However, it did this in python 2, when it should have done it in python 3.

This is now fixed. Attempts at forward campatability are untested and only
there to make future porting easier, nothing is actually intended to work
on Python 3 at the moment.